+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/ContextFileSource.kt
@@ -1,0 +1,5 @@
+@file:Suppress("FunctionName", "ClassName")
+package com.sourcegraph.cody.protocol_generated
+
+typealias ContextFileSource = String // One of: embeddings, user, keyword, editor, filename, search, unified, selection, terminal
+