+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/Date.kt
@@ -1,0 +1,5 @@
+@file:Suppress("FunctionName", "ClassName")
+package com.sourcegraph.cody.protocol_generated
+
+typealias Date = String
+