+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/DeleteOptionsParams.kt
@@ -1,0 +1,8 @@
+@file:Suppress("FunctionName", "ClassName")
+package com.sourcegraph.cody.protocol_generated
+
+data class DeleteOptionsParams(
+  var recursive: Boolean? = null,
+  var ignoreIfNotExists: Boolean? = null,
+)
+