+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/OptionsParams.kt
@@ -1,0 +1,8 @@
+@file:Suppress("FunctionName", "ClassName")
+package com.sourcegraph.cody.protocol_generated
+
+data class OptionsParams(
+  var undoStopBefore: Boolean? = null,
+  var undoStopAfter: Boolean? = null,
+)
+