+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/ProtocolCommand.kt
@@ -1,0 +1,10 @@
+@file:Suppress("FunctionName", "ClassName")
+package com.sourcegraph.cody.protocol_generated
+
+data class ProtocolCommand(
+  var title: String? = null,
+  var command: String? = null,
+  var tooltip: String? = null,
+  var arguments: List<Any>? = null,
+)
+