+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/Selection.kt
@@ -1,0 +1,9 @@
+@file:Suppress("FunctionName", "ClassName")
+package com.sourcegraph.cody.protocol_generated
+
+data class Selection(
+  var anchor: Position? = null,
+  var active: Position? = null,
+  var isReversed: Boolean? = null,
+)
+