+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/UntitledTextDocument.kt
@@ -1,0 +1,9 @@
+@file:Suppress("FunctionName", "ClassName")
+package com.sourcegraph.cody.protocol_generated
+
+data class UntitledTextDocument(
+  var uri: String? = null,
+  var content: String? = null,
+  var language: String? = null,
+)
+