+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/WorkspaceEditEntryMetadata.kt
@@ -1,0 +1,10 @@
+@file:Suppress("FunctionName", "ClassName")
+package com.sourcegraph.cody.protocol_generated
+
+data class WorkspaceEditEntryMetadata(
+  var needsConfirmation: Boolean? = null,
+  var label: String? = null,
+  var description: String? = null,
+  var iconPath: Uri? = null,
+)
+