+++ vscode/src/chat/chat-view/SimpleChatPanelProvider.ts
@@ -547,7 +547,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
             const tabs = getOpenTabsContextFile()
             void this.postMessage({
                 type: 'userContextFiles',
-                context: tabs,
+                userContextFiles: tabs,
             })
             return
         }
@@ -576,7 +576,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
                 if (!cancellation.token.isCancellationRequested) {
                     await this.postMessage({
                         type: 'userContextFiles',
-                        context: symbolResults,
+                        userContextFiles: symbolResults,
                     })
                 }
             } else {
@@ -588,7 +588,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
                 if (!cancellation.token.isCancellationRequested) {
                     await this.postMessage({
                         type: 'userContextFiles',
-                        context: fileResults,
+                        userContextFiles: fileResults,
                     })
                 }
             }
@@ -733,7 +733,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
         )
         void this.postMessage({
             type: 'enhanced-context',
-            context: {
+            enhancedContextStatus: {
                 groups: this.contextStatusAggregator.status,
             },
         })
@@ -1006,7 +1006,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
         if (allHistory) {
             void this.postMessage({
                 type: 'history',
-                messages: allHistory,
+                localHistory: allHistory,
             })
         }
         await this.treeView.updateTree(this.authProvider.getAuthStatus())
@@ -1163,7 +1163,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, ChatSession {
 
         await this.postMessage({
             type: 'view',
-            messages: view,
+            view: view,
         })
     }